indexer/requirements.txt
psycopg2-binary==2.9.10
openai>=1.40.0
beautifulsoup4==4.12.3
numpy>=2.1.0
tqdm==4.66.4
python-dotenv==1.0.1

